fix(falcon-logger): Adopting latest changes from falcon-scripts
