+++ examples/shop-with-blog/server/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@deity/falcon-blog-extension": "^0.1.6",
+    "@deity/falcon-logger": "^0.1.2",
     "@deity/falcon-magento2-api": "^0.6.0",
     "@deity/falcon-server": "^0.2.2",
     "@deity/falcon-shop-extension": "^0.4.2",
